+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [0.0.11] - 2024-08-22 
+
+### Changed
+
+- Updated version to 0.0.11
 
 ### Fixed
 
@@ -14,7 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - checking if no commits
 
 
-## [0.0.10] - 2024-08-22 
+## [0.0.10] - 2024-08-22 [YANKED]
 
 ### Added
 
@@ -183,6 +187,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 [0.0.1]: https://github.com/KDesp73/changelogger/releases/tag/v0.0.1
 [0.0.10]: https://github.com/KDesp73/changelogger/releases/tag/v0.0.10
+[0.0.11]: https://github.com/KDesp73/changelogger/releases/tag/v0.0.11
 [0.0.2]: https://github.com/KDesp73/changelogger/releases/tag/v0.0.2
 [0.0.3]: https://github.com/KDesp73/changelogger/releases/tag/v0.0.3
 [0.0.4]: https://github.com/KDesp73/changelogger/releases/tag/v0.0.4